Wifi heatmap/requirements.txt
Libraries Used:
opencv-python
numpy
pillow
scipy
